requirements.txt
pyyaml
apscheduler
bs4
lxml
requests
selenium
numpy==1.24.3
pillow
opencv-python
packaging

# 识别文字
torch==1.13.1
easyocr